Include modified pdb files in nuget package that points to source on github
This uses GitLink, see https://github.com/GitTools/GitLink to modify the
pdbs to point to the commit on github.

No need to publish a symbols package.

When using GitLink, the user no longer has to specify symbol servers.
It's enough to enable support for source servers in Visual Studio, see
https://github.com/GitTools/GitLink/blob/develop/doc/images/visualstudio_enablesourceserversupport.png

As GitLink 2.1.0 (the latest stable version) requires Debugging Tools
for windows to be installed, i used thhe prerelease version 2.2.0 instead.
That's why -Prerelease is specified when installing nuget packages in
build.ps1