Add Name to Flow, that by default is set to LogType.
Allow Flows with same LogType as long as their names differ.

The existing version did not allow many Flows for the same LogType, but
there is nothing in the code that depends on unique LogTypes. And as it's
possible to deploy many services on the same machine that writes to
the same LogType, it should be possible to do so within a service instance
as well.

To be able to differentiate between Flows, a Name property has been
added that by defaults to the specified LogType. This was done in order to be
backwards compatible.

To add a Flow for the same LogType as a previous flow, a different name must
be specified on the second Flow.